// src/interfaces/ipinfo.ts
export interface IPInfo {
  ip: string;
  asn: string;
  as_org: string | unknown;
  country: string | unknown;
  region: string | unknown;
  city: string | unknown;
  latitude: string | unknown;
  longitude: string | unknown;
  timezone: string | unknown;
  encoding: string | unknown;
  user_agent: string;
}